Split ScheduledCaseInformationUpdaterIT into more appropriate sized test files.
